api/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
orjson==3.10.7

